refactor: Move out webuiRunCmd into own source file
